backend/requirements.txt
fastapi
uvicorn
google-genai  # or the specific package name for the Gemini SDK if different
python-multipart